bank_churn_1/requirements.txt
# Core ML
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Config
pyyaml>=6.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Experiment tracking (optional)
wandb>=0.15.0

# Load testing (optional, dev only)
locust>=2.0.0
